test: drop debug vs release source-set tests
