+++ app/src/test/java/to/bitkit/ui/utils/ScreenDeepLinksTest.kt
@@ -1,6 +1,5 @@
 package to.bitkit.ui.utils
 
-import android.content.Intent
 import android.net.Uri
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -8,6 +7,7 @@ import org.robolectric.RobolectricTestRunner
 import org.robolectric.annotation.Config
 import to.bitkit.test.BaseUnitTest
 import to.bitkit.ui.Routes
+import kotlin.reflect.KClass
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
 import kotlin.test.assertNotNull
@@ -17,6 +17,26 @@ import kotlin.test.assertTrue
 @Config(sdk = [34])
 @RunWith(RobolectricTestRunner::class)
 class ScreenDeepLinksTest : BaseUnitTest() {
+    private companion object {
+        val SENSITIVE_ROUTES: List<KClass<out Routes>> = listOf(
+            Routes.AuthCheck::class,
+            Routes.CriticalUpdate::class,
+            Routes.ExternalAmount::class,
+            Routes.ExternalConfirm::class,
+            Routes.ExternalSuccess::class,
+            Routes.LegacyRnRecovery::class,
+            Routes.LnurlChannel::class,
+            Routes.RecoveryMnemonic::class,
+            Routes.RecoveryMode::class,
+            Routes.SavingsProgress::class,
+            Routes.SettingUp::class,
+            Routes.SpendingAdvanced::class,
+            Routes.SpendingConfirm::class,
+            Routes.SpendingHwSign::class,
+            Routes.SpendingHwSigned::class,
+        )
+    }
+
     @Test
     fun `screen id is derived from the route name in kebab-case`() {
         val home = ScreenDeepLinks.screenId(Routes.Home::class)
@@ -29,6 +49,38 @@ class ScreenDeepLinksTest : BaseUnitTest() {
     }
 
     @Test
+    fun `routes without arguments produce a bare pattern`() {
+        val links = ScreenDeepLinks.linksFor(Routes.Settings::class)
+
+        assertEquals(1, links.size)
+        assertEquals("bitkit://screen/settings", links.single().uriPattern)
+    }
+
+    @Test
+    fun `required arguments are appended as path segments`() {
+        val links = ScreenDeepLinks.linksFor(Routes.ActivityAssignContact::class)
+
+        assertEquals("bitkit://screen/activity-assign-contact/{id}", links.single().uriPattern)
+    }
+
+    @Test
+    fun `a route with both argument kinds keeps the required one in the path`() {
+        val links = ScreenDeepLinks.linksFor(Routes.ActivityDetail::class)
+
+        assertEquals("bitkit://screen/activity-detail/{id}?walletId={walletId}", links.single().uriPattern)
+    }
+
+    @Test
+    fun `arguments with defaults are appended as query parameters`() {
+        val links = ScreenDeepLinks.linksFor(Routes.Contacts::class)
+
+        assertEquals(
+            "bitkit://screen/contacts?showAddContactSheet={showAddContactSheet}",
+            links.single().uriPattern,
+        )
+    }
+
+    @Test
     fun `every route declares whether it may be entered directly`() {
         val markers = setOf(Routes.DeepLinkable::class, Routes.InternalOnly::class)
 
@@ -38,7 +90,7 @@ class ScreenDeepLinksTest : BaseUnitTest() {
     }
 
     @Test
-    fun `every deep-linkable route has a unique screen id and variant-correct links`() {
+    fun `every deep-linkable route has a unique screen id and one link`() {
         val ids = mutableMapOf<String, String>()
 
         Routes.DeepLinkable::class.sealedSubclasses.forEach { route ->
@@ -48,12 +100,7 @@ class ScreenDeepLinksTest : BaseUnitTest() {
             assertNotNull(id, "route $name has no screen id")
             val clash = ids.put(id, name)
             assertNull(clash, "screen id '$id' is used by both $clash and $name")
-            val links = ScreenDeepLinks.linksFor(route)
-            if (ScreenDeepLinks.isEnabled) {
-                assertEquals(1, links.size, "route $name has no deep link")
-            } else {
-                assertTrue(links.isEmpty(), "route $name leaked a deep link in release")
-            }
+            assertEquals(1, ScreenDeepLinks.linksFor(route).size, "route $name has no deep link")
         }
     }
 
@@ -84,20 +131,5 @@ class ScreenDeepLinksTest : BaseUnitTest() {
         assertFalse(recoveryMode)
         assertFalse(pubkyAuth)
         assertFalse(lightning)
-    }
-
-    @Test
-    fun `screen links are enabled only on debug`() {
-        assertEquals(ScreenDeepLinks.isEnabled, ScreenDeepLinks.linksFor(Routes.Settings::class).isNotEmpty())
-    }
-
-    @Test
-    fun `screen uris are detached from the activity intent`() {
-        val intent = Intent(Intent.ACTION_VIEW, Uri.parse("bitkit://screen/settings"))
-
-        val detached = ScreenDeepLinks.detachScreenUri(intent)
-
-        assertTrue(detached)
-        assertNull(intent.data)
     }
 }
